mnist digit recognition/requirements.txt
gradio
scikit-learn==1.7.1
joblib
numpy
pillow
